requirements.txt
# requirements-core.txt — Dépendances obligatoires (API + Desktop)
altgraph==0.17.5
asn1crypto==1.5.1
certifi==2026.2.25
certvalidator==0.11.1
cffi==2.0.0
charset-normalizer==3.4.6
colorama==0.4.6
cryptography==46.0.5
customtkinter==5.2.2
darkdetect==0.8.0
fonttools==4.62.1
idna==3.11
lxml==6.0.2
numpy==2.4.3
opencv-python-headless==4.13.0.92
oscrypto==1.3.0
packaging==26.0
pefile==2024.8.26
pillow==12.1.1
pycparser==3.0
pyHanko==0.34.1
pyhanko-certvalidator==0.30.1
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
PyMuPDF==1.27.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
qrcode==8.2
requests==2.32.5
setuptools==82.0.1
tzdata==2025.3
tzlocal==5.3.1
uharfbuzz==0.53.3
uritools==6.0.1
urllib3==2.6.3

# ── Sécurité (nouvelles dépendances) ──
bcrypt>=4.2
PyJWT>=2.8

# ── API Dependencies ──
fastapi
uvicorn
python-multipart
pydantic
slowapi

# ── Worker Dependencies ──
celery
redis

# ── Env management ──
python-dotenv>=1.0
